requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
dnspython==2.3.0
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.89.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
jose==1.0.0
mysqlclient==2.1.1
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
PyMySQL==1.0.2
pytest==7.2.1
python-dotenv==0.21.0
python-multipart==0.0.5
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
